Add support for post processors
